requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1
sqlalchemy==2.0.36
alembic==1.14.1
aiosqlite==0.20.0
arq==0.26.1
redis==5.2.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
bcrypt==4.2.1
python-multipart==0.0.20
websockets==14.2
